docker update, dot env support
