+++ py_fmg/api/main.py
@@ -3,6 +3,7 @@
 from fastapi import FastAPI, BackgroundTasks, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel, Field
+from typing import List, Optional, Tuple
 from typing import Optional
 import structlog
 import uuid
