+++ py_fmg/config/__init__.py
@@ -3,5 +3,6 @@ Configuration modules for map generation.
 """
 
 from .heightmap_templates import get_template, list_templates, TEMPLATES
+from .config import settings
 
-__all__ = ['get_template', 'list_templates', 'TEMPLATES']
+__all__ = ['get_template', 'list_templates', 'TEMPLATES', "settings"]
