+++ py_fmg/db/connection.py
@@ -1,6 +1,6 @@
 """Database connection utilities."""
 
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, text  # ✅ Added text here
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 import structlog
@@ -14,52 +14,52 @@ logger = structlog.get_logger()
 
 class Database:
     """Database connection manager."""
-    
+
     def __init__(self):
         self.engine = None
         self.SessionLocal = None
-        
+
     def initialize(self):
         """Initialize database connection."""
         logger.info("Initializing database connection", url=settings.database_url)
-        
+
         # Create engine
         self.engine = create_engine(
             settings.database_url,
             poolclass=StaticPool,
             echo=False,  # Set to True for SQL debugging
         )
-        
+
         # Create session factory
         self.SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=self.engine)
-        
+
         # Create tables
         self.create_tables()
-        
+
         logger.info("Database connection initialized")
-    
+
     def create_tables(self):
         """Create all tables."""
         try:
             # Ensure PostGIS extension exists
             with self.engine.connect() as conn:
-                conn.execute("CREATE EXTENSION IF NOT EXISTS postgis")
+                conn.execute(text("CREATE EXTENSION IF NOT EXISTS postgis"))  # ✅ Wrapped in text()
                 conn.commit()
-            
+
             # Create all tables
             Base.metadata.create_all(bind=self.engine)
             logger.info("Database tables created")
-            
+
         except Exception as e:
             logger.error("Failed to create tables", error=str(e))
             raise
-    
+
     @contextmanager
     def get_session(self):
         """Get database session with automatic cleanup."""
         if not self.SessionLocal:
             raise RuntimeError("Database not initialized. Call initialize() first.")
-        
+
         session = self.SessionLocal()
         try:
             yield session
